requirements.txt
django==3.2.12
djangorestframework==3.13.1
urllib3==1.26.9
xsdata==22.5
lxml==4.9.1